🗄 Reorganized Profile Components
