+++ react-app/src/Components/Profile/index.js
@@ -1,0 +1,1 @@
+export { default as Profile } from './Profile'
